backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.12
alembic==1.13.3
redis==5.1.1
celery==5.4.0
aiohttp==3.10.8
boto3==1.35.36
minio==7.2.9
cryptography==43.0.3
httpx==0.27.2
python-multipart==0.0.12
yt-dlp==2024.10.7
openai==1.51.2
google-generativeai==0.8.3
tenacity==9.0.0
orjson==3.10.7
slowapi==0.1.9
websockets==13.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0

